# scalar quantization: add SIMD dot product