--- a/adapters/repos/db/vector/hnsw/distancer/dot_product_amd64_test.go
+++ b/adapters/repos/db/vector/hnsw/distancer/dot_product_amd64_test.go
@@ -31,6 +31,16 @@ var dotByteImpl func(a, b []byte) uint32 = func(a, b []byte) uint32 {
 	return uint32(sum)
 }
 
+var dotFloatByteImpl func(a []float32, b []byte) float32 = func(a []float32, b []byte) float32 {
+	var sum float32
+
+	for i := range a {
+		sum += a[i] * float32(b[i])
+	}
+
+	return sum
+}
+
 func testDotProductFixedValue(t *testing.T, size uint, dotFn func(x []float32, y []float32) float32) {
 	count := 10000
 	countFailed := 0
@@ -257,6 +267,121 @@ func BenchmarkDotByte(b *testing.B) {
 
 			// b.Run("pure go", func(b *testing.B) { benchmarkDotByte(b, dim, dotByteImpl) })
 			// b.Run("avx", func(b *testing.B) { benchmarkDotByte(b, dim, asm.DotByteAVX256) })
+		})
+	}
+}
+
+func testDotProductFloatByteFixedValue(t *testing.T, size uint, dotFn func(x []float32, y []uint8) float32) {
+	vec1 := make([]float32, size)
+	vec2 := make([]uint8, size)
+	for i := range vec1 {
+		vec1[i] = 1
+		vec2[i] = 1
+	}
+	res := dotFn(vec1, vec2)
+
+	resControl := dotFloatByteImpl(vec1, vec2)
+	if resControl != res {
+		t.Logf("for dim: %d -> want: %f, got: %f", size, resControl, res)
+		t.Fail()
+	}
+}
+
+func testDotProductFloatByteRandomValue(t *testing.T, size uint, dotFn func(x []float32, y []byte) float32) {
+	r := getRandomSeed()
+	count := 10000
+
+	vec1s := make([][]float32, count)
+	vec2s := make([][]byte, count)
+
+	for i := 0; i < count; i++ {
+		vec1 := make([]float32, size)
+		vec2 := make([]byte, size)
+		for j := range vec1 {
+			vec1[j] = float32(r.Uint32() % 1000)
+			vec2[j] = byte(r.Uint32() % 256)
+		}
+
+		vec1s[i] = Normalize(vec1)
+		vec2s[i] = vec2
+	}
+
+	for i := 0; i < count; i++ {
+		res := dotFn(vec1s[i], vec2s[i])
+
+		resControl := dotFloatByteImpl(vec1s[i], vec2s[i])
+		delta := float64(0.05)
+		diff := float64(resControl) - float64(res)
+		if diff < -delta || diff > delta {
+			t.Logf("for dim: %d -> want: %f, got: %f, diff: %f", size, resControl, res, diff)
+			t.Fail()
+		}
+	}
+}
+
+func TestCompareDotProductFloatByte(t *testing.T) {
+	sizes := []uint{
+		1,
+		2,
+		3,
+		4,
+		5,
+		6,
+		8,
+		10,
+		12,
+		16,
+		24,
+		30,
+		31,
+		32,
+		64,
+		67,
+		128,
+		256,
+		260,
+		299,
+		300,
+		384,
+		390,
+		600,
+		768,
+		777,
+		784,
+		1024,
+		1536,
+	}
+
+	for _, size := range sizes {
+		t.Run(fmt.Sprintf("with size %d", size), func(t *testing.T) {
+			testDotProductFloatByteFixedValue(t, size, asm.DotFloatByteAVX256)
+			testDotProductFloatByteRandomValue(t, size, asm.DotFloatByteAVX256)
+		})
+	}
+}
+
+func benchmarkDotFloatByte(b *testing.B, dims int, dotFn func(a []float32, b []byte) float32) {
+	r := getRandomSeed()
+
+	vec1 := make([]float32, dims)
+	vec2 := make([]byte, dims)
+	for i := range vec1 {
+		vec1[i] = float32(r.Uint32() % 1000)
+		vec2[i] = byte(r.Uint32() % 256)
+	}
+
+	b.ResetTimer()
+	for n := 0; n < b.N; n++ {
+		dotFn(vec1, vec2)
+	}
+}
+
+func BenchmarkDotFloatByte(b *testing.B) {
+	dims := []int{2, 4, 6, 8, 10, 12, 16, 24, 30, 32, 128, 256, 300, 384, 512, 768, 1024, 1536}
+	for _, dim := range dims {
+		b.Run(fmt.Sprintf("%d dimensions", dim), func(b *testing.B) {
+			b.Run("pure go", func(b *testing.B) { benchmarkDotFloatByte(b, dim, dotFloatByteImpl) })
+			b.Run("avx", func(b *testing.B) { benchmarkDotFloatByte(b, dim, asm.DotFloatByteAVX256) })
 		})
 	}
 }