### What's being changed:

```
benchstat sq_dot_go.txt sq_dot_avx2.txt
goos: linux
goarch: amd64
pkg: github.com/weaviate/weaviate/adapters/repos/db/vector/hnsw/distancer
cpu: Intel(R) Xeon(R) CPU @ 2.60GHz
                               │ sq_dot_go.txt │           sq_dot_avx2.txt           │
                               │    sec/op     │    sec/op     vs base               │
DotFloatByte/2_dimensions-4        3.501n ± 0%   5.043n ±  2%  +44.08% (p=0.002 n=6)
DotFloatByte/4_dimensions-4        5.223n ± 0%   5.716n ±  1%   +9.46% (p=0.002 n=6)
DotFloatByte/6_dimensions-4        6.184n ± 7%   7.350n ±  0%  +18.85% (p=0.002 n=6)
DotFloatByte/8_dimensions-4        6.559n ± 1%   5.736n ±  1%  -12.55% (p=0.002 n=6)
DotFloatByte/10_dimensions-4       7.795n ± 0%   6.855n ±  1%  -12.05% (p=0.002 n=6)
DotFloatByte/12_dimensions-4       8.977n ± 0%   7.439n ±  0%  -17.13% (p=0.002 n=6)
DotFloatByte/16_dimensions-4      11.995n ± 0%   6.538n ±  0%  -45.49% (p=0.002 n=6)
DotFloatByte/24_dimensions-4      18.215n ± 1%   6.936n ±  1%  -61.92% (p=0.002 n=6)
DotFloatByte/30_dimensions-4      24.070n ± 0%   9.500n ±  0%  -60.53% (p=0.002 n=6)
DotFloatByte/32_dimensions-4      26.385n ± 1%   6.863n ±  0%  -73.99% (p=0.002 n=6)
DotFloatByte/128_dimensions-4     128.10n ± 0%   11.39n ±  4%  -91.11% (p=0.002 n=6)
DotFloatByte/256_dimensions-4     285.60n ± 0%   17.48n ± 23%  -93.88% (p=0.002 n=6)
DotFloatByte/300_dimensions-4     337.80n ± 0%   21.50n ±  1%  -93.64% (p=0.002 n=6)
DotFloatByte/384_dimensions-4     437.25n ± 0%   27.48n ±  1%  -93.71% (p=0.002 n=6)
DotFloatByte/512_dimensions-4     588.65n ± 0%   33.14n ±  2%  -94.37% (p=0.002 n=6)
DotFloatByte/768_dimensions-4     892.55n ± 2%   44.77n ±  2%  -94.98% (p=0.002 n=6)
DotFloatByte/1024_dimensions-4   1196.00n ± 0%   56.77n ±  1%  -95.25% (p=0.002 n=6)
DotFloatByte/1536_dimensions-4   1804.50n ± 0%   77.38n ±  1%  -95.71% (p=0.002 n=6)
geomean                            57.28n        13.16n        -77.02%
```

### Review checklist

- [ ] Documentation has been updated, if necessary. Link to changed documentation:
- [ ] Chaos pipeline run or not necessary. Link to pipeline:
- [ ] All new code is covered by tests where it is reasonable.
- [ ] Performance tests have been run or not necessary.